Move request sessions and flusher to `ServerRuntimeClient`
