+++ packages/node/src/client.ts
@@ -1,7 +1,5 @@
-import type { Scope, ServerRuntimeClientOptions } from '@sentry/core';
-import { SDK_VERSION, ServerRuntimeClient, SessionFlusher } from '@sentry/core';
-import type { Event, EventHint } from '@sentry/types';
-import { logger } from '@sentry/utils';
+import type { ServerRuntimeClientOptions } from '@sentry/core';
+import { SDK_VERSION, ServerRuntimeClient } from '@sentry/core';
 import * as os from 'os';
 import { TextEncoder } from 'util';
 
@@ -14,8 +12,6 @@ import type { NodeClientOptions } from './types';
  * @see SentryClient for usage documentation.
  */
 export class NodeClient extends ServerRuntimeClient<NodeClientOptions> {
-  protected _sessionFlusher: SessionFlusher | undefined;
-
   /**
    * Creates a new Node SDK instance.
    * @param options Configuration options for this SDK.
@@ -47,87 +43,5 @@ export class NodeClient extends ServerRuntimeClient<NodeClientOptions> {
     };
 
     super(clientOptions);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
-  public captureException(exception: any, hint?: EventHint, scope?: Scope): string | undefined {
-    // Check if the flag `autoSessionTracking` is enabled, and if `_sessionFlusher` exists because it is initialised only
-    // when the `requestHandler` middleware is used, and hence the expectation is to have SessionAggregates payload
-    // sent to the Server only when the `requestHandler` middleware is used
-    if (this._options.autoSessionTracking && this._sessionFlusher && scope) {
-      const requestSession = scope.getRequestSession();
-
-      // Necessary checks to ensure this is code block is executed only within a request
-      // Should override the status only if `requestSession.status` is `Ok`, which is its initial stage
-      if (requestSession && requestSession.status === 'ok') {
-        requestSession.status = 'errored';
-      }
-    }
-
-    return super.captureException(exception, hint, scope);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public captureEvent(event: Event, hint?: EventHint, scope?: Scope): string | undefined {
-    // Check if the flag `autoSessionTracking` is enabled, and if `_sessionFlusher` exists because it is initialised only
-    // when the `requestHandler` middleware is used, and hence the expectation is to have SessionAggregates payload
-    // sent to the Server only when the `requestHandler` middleware is used
-    if (this._options.autoSessionTracking && this._sessionFlusher && scope) {
-      const eventType = event.type || 'exception';
-      const isException =
-        eventType === 'exception' && event.exception && event.exception.values && event.exception.values.length > 0;
-
-      // If the event is of type Exception, then a request session should be captured
-      if (isException) {
-        const requestSession = scope.getRequestSession();
-
-        // Ensure that this is happening within the bounds of a request, and make sure not to override
-        // Session Status if Errored / Crashed
-        if (requestSession && requestSession.status === 'ok') {
-          requestSession.status = 'errored';
-        }
-      }
-    }
-
-    return super.captureEvent(event, hint, scope);
-  }
-
-  /**
-   *
-   * @inheritdoc
-   */
-  public close(timeout?: number): PromiseLike<boolean> {
-    this._sessionFlusher?.close();
-    return super.close(timeout);
-  }
-
-  /** Method that initialises an instance of SessionFlusher on Client */
-  public initSessionFlusher(): void {
-    const { release, environment } = this._options;
-    if (!release) {
-      __DEBUG_BUILD__ && logger.warn('Cannot initialise an instance of SessionFlusher if no release is provided!');
-    } else {
-      this._sessionFlusher = new SessionFlusher(this, {
-        release,
-        environment,
-      });
-    }
-  }
-
-  /**
-   * Method responsible for capturing/ending a request session by calling `incrementSessionStatusCount` to increment
-   * appropriate session aggregates bucket
-   */
-  protected _captureRequestSession(): void {
-    if (!this._sessionFlusher) {
-      __DEBUG_BUILD__ && logger.warn('Discarded request mode session because autoSessionTracking option was disabled');
-    } else {
-      this._sessionFlusher.incrementSessionStatusCount();
-    }
   }
 }
